2nd round of de-linting
